Move CliSearchAction to correct package
